requirements.txt
streamlit
pandas
openpyxl
xlrd
odfpy
pdfplumber
google-generative-ai
orjson
tqdm
python-dotenv